router_project/requirements.txt
fastapi==0.110.*
uvicorn==0.29.*
groq==0.5.*
python-dotenv==1.0.*
sqlalchemy==2.0.*
pydantic==2.6.*
